refactor(all):
update examples, fix dependencies, update docs in readme.rst, some types are tested by hands and actualized, implement new exception handling, implement experimental htstring, implement connector, apply formatters [isort, black], add new return type, add new ToolSet class to core/requests.py

+++ aioqiwi/wallet/types/commission.py
@@ -9,7 +9,7 @@ from aioqiwi.types import BaseModel
 class FundingSourceCommission(BaseModel):
     """Object: fundingSourceCommission"""
 
-    amount: int = Field(..., alias="amount")
+    amount: float = Field(..., alias="amount")
     currency: str = Field(..., alias="currency")
 
 
@@ -23,7 +23,7 @@ class QwCommission(BaseModel):
 class EnrollmentSum(BaseModel):
     """Object: enrollmentSum"""
 
-    amount: int = Field(..., alias="amount")
+    amount: float = Field(..., alias="amount")
     currency: str = Field(..., alias="currency")
